+++ linkerd/http-retry/Cargo.toml
@@ -12,11 +12,8 @@ futures = { version = "0.3", default-features = false }
 http-body = "0.4"
 http = "0.2"
 linkerd-error = { path = "../error" }
-linkerd-http-box = { path = "../http-box" }
-linkerd-stack = { path = "../stack" }
 pin-project = "1"
 parking_lot = "0.11"
-tower = { version = "0.4.7", default-features = false, features = ["retry", "util"] }
 tracing = "0.1.23"
 thiserror = "1"
